App/App_Test_1/flask_application/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.6.2
asttokens==2.0.5
astunparse==1.6.3
bottleneck==1.4.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.112.2
flatbuffers==25.2.10
fonttools==4.55.3
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
keras==2.15.0
kiwisolver==1.4.8
lazy-loader==0.4
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pip==25.0
platformdirs==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.3
pydantic-core==2.27.1
pygments==2.15.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
sip==6.10.0
six==1.16.0
sniffio==1.3.0
stack-data==0.2.0
starlette==0.38.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.0.1
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.32.1
wcwidth==0.2.5
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
